API/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
motor>=3.5.1
pydantic>=2.7.4
pydantic-settings>=2.1.0
python-dotenv==1.0.0
bcrypt==4.1.2
PyJWT==2.8.0
cryptography==42.0.0
langchain==0.3.14
langchain-openai==0.3.0
langchain-anthropic==0.3.0
langchain-google-genai==2.0.0
tiktoken>=0.5.2
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.12.0
pytest-cov>=4.1.0
mongomock-motor>=0.0.21

